pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "serendipity-spend"
version = "0.1.0"
description = "Expense reimbursement intake and review for travel and related expenses."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.13.2",
  "boto3>=1.34.160",
  "celery>=5.4.0",
  "email-validator>=2.2.0",
  "fastapi>=0.115.0",
  "jinja2>=3.1.4",
  "openpyxl>=3.1.5",
  "passlib[bcrypt]>=1.7.4",
  "psycopg[binary]>=3.2.1",
  "pydantic>=2.8.2",
  "pydantic-settings>=2.4.0",
  "pypdf[image]>=5.0.0",
  "pytesseract>=0.3.10",
  "python-jose[cryptography]>=3.3.0",
  "python-multipart>=0.0.9",
  "redis>=5.0.8",
  "sqlalchemy>=2.0.32",
  "uvicorn[standard]>=0.30.6",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "pytest-cov>=5.0.0",
  "ruff>=0.5.7",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"
exclude = ["alembic/versions"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["B008"]

[tool.pytest.ini_options]
testpaths = ["tests"]
